feat: integrated curated websites list in publishers screen
