+++ src/app/screens/Publishers/websites.json
@@ -1,0 +1,105 @@
+[
+  {
+    "title": "Trading",
+    "items": [
+      {
+        "title": "LNMarkets",
+        "subtitle": "Instant Bitcoin derivatives trading",
+        "logo": "https://testnet.lnmarkets.com/apple-touch-icon.png",
+        "url": "https://testnet.lnmarkets.com/"
+      },
+      {
+        "title": "Kollider",
+        "subtitle": "Instant Bitcoin derivatives trading",
+        "logo": "https://kollider.xyz/favicon0.ico",
+        "url": "https://kollider.xyz/"
+      },
+      {
+        "title": "LOFT",
+        "subtitle": "Trade investment products in Bitcoin",
+        "logo": "https://loft.trade/assets/logo-small.svg",
+        "url": "https://loft.trade/"
+      }
+    ]
+  },
+  {
+    "title": "Gaming",
+    "items": [
+      {
+        "title": "LNGames",
+        "subtitle": "Simple Lightning Network games",
+        "logo": "https://lngames.net/images/newlogo.png",
+        "url": "https://lngames.net/"
+      },
+      {
+        "title": "Lightining Poker",
+        "subtitle": "Play Poker with Bitcoin",
+        "logo": "https://lightning-poker.com/favicon.ico",
+        "url": "https://lightning-poker.com/"
+      },
+      {
+        "title": "LNBlackJack",
+        "subtitle": "Play Blackjack with Bitcoin",
+        "logo": "https://www.lnblackjack.com/images/table.png",
+        "url": "https://www.lnblackjack.com/"
+      },
+      {
+        "title": "Lightning Roulette",
+        "subtitle": "Play Roulette with Bitcoin",
+        "logo": "https://lightning-roulette.com/favicon.ico",
+        "url": "https://lightning-roulette.com/"
+      }
+    ]
+  },
+  {
+    "title": "Entertaiment",
+    "items": [
+      {
+        "title": "Stacker.News",
+        "subtitle": "Lightning powered Bitcoin news site",
+        "logo": "https://stacker.news/favicon.png",
+        "url": "https://stacker.news/"
+      }
+    ]
+  },
+  {
+    "title": "Shopping",
+    "items": [
+      {
+        "title": "Bitrefill",
+        "subtitle": "Buy vouchers, refill your phone or pay your bills",
+        "logo": "https://www.bitrefill.com/favicon.svg",
+        "url": "https://www.bitrefill.com/?hl=en"
+      },
+      {
+        "title": "Lightning Network Stores",
+        "subtitle": "Collection of stores and websites accepting Bitcoin payments",
+        "logo": "https://cointelegraph.com/favicons/apple-touch-icon.png",
+        "url": "https://cointelegraph.com/tags/lightning-network"
+      }
+    ]
+  },
+  {
+    "title": "Miscellaneous",
+    "items": [
+      {
+        "title": "Sparkshot",
+        "subtitle": "Art created and funded by the community",
+        "logo": "https://sparkshot.io/favicon/android-icon-192x192.png",
+        "url": "https://sparkshot.io/"
+      },
+      {
+        "title": "Sats for Likes",
+        "subtitle": "Earn sats for accomplishing tasks",
+        "logo": "https://kriptode.com/satsforlikes/web_hi_res_512.webp",
+        "url": "https://kriptode.com/satsforlikes/index.html"
+      },
+      {
+        "title": "lnshort.it",
+        "subtitle": "Service to redirect URLs and generate QR codes",
+        "logo": "https://lnshort.it/static/lnshort/img/favicon.ico",
+        "url": "https://lnshort.it/"
+      }
+    ]
+  }
+]
